flask/requirements.txt
Flask
gspread
Flask-CORS
Pillow
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
unidecode
Gunicorn
